requirements.txt
absl-py==2.1.0
asgiref==3.3.1
astroid==2.4.2
astunparse==1.6.3
boto3==1.16.47
botocore==1.19.47
cachetools==5.3.2
certifi==2020.12.5
chardet==4.0.0
charset-normalizer==3.3.2
Django==3.1.5
django-extensions==3.1.0
django-filter==2.4.0
django-storages==1.11.1
djangorestframework==3.12.2
flatbuffers==23.5.26
gast==0.5.4
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.0
gunicorn==20.0.4
h5py==3.10.0
idna==2.10
isort==5.6.4
jmespath==0.10.0
keras==2.15.0
lazy-object-proxy==1.4.3
libclang==16.0.6
Markdown==3.3.3
MarkupSafe==2.1.3
mccabe==0.6.1
ml-dtypes==0.2.0
mysqlclient==2.0.3
numpy==1.26.2
oauthlib==3.2.2
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
Pillow==8.0.1
protobuf==4.23.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydotplus==2.0.2
PyJWT==2.0.0
pylint==2.6.0
pyparsing==2.4.7
PyQt5==5.15.10
PyQt5-Qt5==5.15.12
PyQt5-sip==12.13.0
python-dateutil==2.8.1
python-dotenv==1.0.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.15.0
termcolor==2.4.0
toml==0.10.2
twilio==6.50.1
typing_extensions==4.9.0
urllib3==1.26.2
Werkzeug==3.0.1
whitenoise==5.2.0
wrapt==1.12.1
